move pkg/docker to pkg/container
