+++ pkg/container/client.go
@@ -1,4 +1,4 @@
-package docker
+package container
 
 import (
 	"context"
