+++ pkg/container/docker_mocks.go
@@ -1,4 +1,4 @@
-package docker
+package container
 
 import (
 	"context"
