+++ pkg/container/docker_test.go
@@ -1,4 +1,4 @@
-package docker
+package container
 
 import (
 	"bufio"
